refactor(native): extract platform lifecycle
